rename treebuilder to tree_builder to be consistent with assoc_test snakecase format
