+++ happler/tree/__init__.py
@@ -1,5 +1,5 @@
 from .tree import Tree
-from .treebuilder import TreeBuilder
+from .tree_builder import TreeBuilder
 from .variant import VariantType, Variant
 from .haplotypes import Haplotype, Haplotypes
 from .assoc_test import AssocTest, AssocTestSimple
